Document your conditions

Documents your conditions() and conditions()

cnd_document(
  package = get_package(),
  registry = package,
  file = file.path("R", paste0(package, "-cnd-conditions.R")),
  cleanup = TRUE
)

cnd_section(fun)

Arguments

package

The package to document

registry

The name of the registry

file

The file to save the documentation. This can be a file path, a connection object, or NULL. When file is a path, the directory of the path is searched for files containing # % Generated by cnd: do not edit by hand. These are removed if they are not the same as the generated documentation.

cleanup

If FALSE will not remove files containing # % Generated by cnd: do not edit by hand

fun

The name of a function

Value

Conditional on the file argument:

when file is a connection, the connection object

when file is a path, the path

when file is NULL, a character vector of the documentation

if no conditions are found, a warning is thrown and NULL is returned

conditions

Conditions are generated through the {cnd} package. The following conditions are associated with this function:

For more conditions, see: