Conditions

condition() is used to create a new condition function that itself returns a new condition.

conditions() retrieves all conditions based on search values. The parameters serve as filtering arguments.

condition(
  class,
  message = NULL,
  type = c("error", "warning", "message", "condition"),
  package = get_package(),
  exports = NULL,
  help = NULL,
  registry = package,
  register = !is.null(registry)
)

conditions(
  ...,
  class = NULL,
  type = NULL,
  package = NULL,
  registry = NULL,
  fun = NULL
)

cond(x)

cnd(condition)

conditions(x, ...) <- value

# S3 method for class '`function`'
conditions(x, append = FALSE, ...) <- value

# S3 method for class '`cnd::condition_progenitor`'
conditions(x, ...) <- value

Arguments

class

The name of the new class

message

The message to be displayed when the condition is called

type

The type of condition: error, warning, or message

package

The package to which the condition belongs

exports

The exported functions to be displayed when the condition is called

help

The help message to be displayed for the condition function

registry

The name of the registry to store the condition

register

Controls registration checks

...

Additional arguments passed to methods

fun

if a function is passed, then retrieves the "conditions" attribute

x

An object

condition

A condition object

value

A condition

append

If TRUE, adds to the list of conditions

Value

condition() a condition_generator object

conditions() a list of all conditions

cnd() is a wrapper for calling stop(), warning(), or message(); when condition is a type, an error is thrown, and likewise for the other types. When an error isn't thrown, the condition is returned, invisibly.

Details

Conditions

condition_generator

A condition_generator is an object (a special function) which can be used to create generate a new condition, based on specifications applied in condition(). These functions use ... to absorb extra arguments and contain a special .call parameter. By default, .call captures the parent call from where the condition_generator was created, but users may pass their own call to override this. See call. in conditionCall()

condition() conditions

Conditions are generated through the {cnd} package. The following conditions are associated with this function:

For more conditions, see:

cnd() conditions

Conditions are generated through the {cnd} package. The following conditions are associated with this function:

cnd:as_character_cnd_error/error

You cannot coerce a condition_generator object to a character. This may have occured when trying to put a condition function through stop() or warning. Instead, call the function first, then pass the result to stop() or warning().

For example:

# Instead of this
stop(my_condition)

# Do this
stop(my_condition())

cnd:cnd_document_conditions/warning

cnd:cnd_document_file/error

cnd:cnd_document_pkg_reg/error

cnd:cnd_generated_cleanup/message

cnd:cnd_generated_write/condition

cnd:cond_cnd_class/error

cnd() simple calls the appropriate function: stop(), warning(), or message() based on the type parameter from condition().

cnd:condition_message_generator/error

'cnd::condition_generator' objects are not conditions. You may have made this mistake: rx <- condition('my_condition')conditionMessage(x)Condition generators need to be called first before they can be used as conditions. Try this instead:rx <- condition('my_condition')conditionMessage(x())

cnd:condition_overwrite/warning

cnd:conditions_dots/warning

The ... parameter in conditions() is meant for convenience. Onlya single argument is alowed. Other parameters must be named explicitly.For example:r# Instead of thisconditions('class', 'package') # 'package' is ignored with a warning# Do thisconditions(class = 'class', package = 'package')...

cnd:invalid_condition/error

The class, exports, and help parameters must be a single character string. If you are passing a function, it must be a valid function.

cnd:invalid_condition_message/error

Conditions messages are displayed when invoked through conditionMessage(). You can set a static message by passing through a character vector, or a dynamic message by passing through a function. The function should return a character vector.

When message is not set, a default "there was an error" message is used.

cnd:match_arg/error

Mostly match.arg() but with a custom condition

cnd:no_package_exports/warning

The exports parameter requires a package

For more conditions, see: cnd-cnd-conditions

See also

Examples

# create a new condition:
cond_bad_value <- condition("bad_value")

# use the condition
try(stop(cond_bad_value()))
#> Error in eval(expr, envir) : <bad_value>
#>  Error in eval(expr, envir) : there was an error
try(cnd(cond_bad_value()))
#> Error in eval(expr, envir) : <bad_value>
#>  Error in eval(expr, envir) : there was an error

# dynamic messages:
cond_class_error <- condition(
  "class_error",
  message = function(x) paste("class cannot be", toString(class(x)))
)
try(stop(cond_class_error(list())))
#> Error in eval(expr, envir) : <class_error>
#>  Error in eval(expr, envir) : class cannot be list